Move Linux-only CLI args and wl_interpose anchor out of the entry point
The Linux CLI parsing moves into jfn-linux-util and the wl_interpose
ensure_linked() call moves into platform_install's install_early, so
main.rs and cli.rs carry no platform conditional. The interpose module
is now private to the crate.
